requirements.txt
Django>=1.8,<2.0
psycopg2
django-admin
social-auth-app-django
django-localflavor